Muda a assinatura de Clamp e correções de avisos

+++ framework/platform-dx/rasterizerstate.cpp
@@ -68,7 +68,7 @@ namespace xna {
 	}
 
 	float RasterizerState::DepthBias() const {
-		return impl->dxDescription.DepthBias;
+		return static_cast<float>(impl->dxDescription.DepthBias);
 	}
 
 	void RasterizerState::DepthBias(float value) {
